+++ src/pages/reports/reportsDetail.js
@@ -1,11 +1,7 @@
 import React from 'react';
-import { connect } from "react-redux";
-import { compose } from "redux";
-import { firebaseConnect } from "react-redux-firebase";
 import {List, Divider } from 'antd';
 import { Link } from "react-router-dom";
 import moment from 'moment';
-import Reports from './reports.component';
 import {groupBy} from '../../utils';
 
 window.moment = moment;
@@ -15,11 +11,8 @@ const getD = (data, filter) =>  data.filter(item => {
     return filter.month === month && filter.year == year;
 });
 
-const ReportsDetail = (props) => {
-    console.log('props', props);
+const ReportsDetail = ({ filterData, group, ...props }) => {
     if (props.data && props.data.length) {
-        const filterData = getD(props.data, {year:props.match.params.year, month: props.match.params.month});
-        const group = groupBy(filterData, item => item.value.category);
         let savings =  filterData.filter(item => item.value.category === 'Savings');
         savings = savings.reduce((a,b) => ({value: {amount: parseFloat(a.value.amount)+parseFloat(b.value.amount)}})).value.amount;
         const total = filterData.reduce((a,b) => ({value: {amount: parseFloat(a.value.amount)+parseFloat(b.value.amount)}})).value.amount;
@@ -42,7 +35,6 @@ const ReportsDetail = (props) => {
             <Divider />
             <h1>Reports {props.match.params.month} {props.match.params.year}</h1>
             <Divider />
-            
             <List
                 bordered
                 dataSource={stats}
@@ -91,30 +83,5 @@ const ReportsDetail = (props) => {
     }
 }
 
-const enhancer =  compose(
-    firebaseConnect(
-        props => (
-            [
-                {
-                    path: `data/${props.uid}`,
-                    storeAs: 'data',
-                    queryParams: ['orderByChild=date']
-                }
-            ]
-        )),
-    connect(
-        ({firebase}) => ({
-            data: firebase.ordered.data,
-            uid: firebase.auth.uid
-        }))
-);
 
-
-const SummaryEnhancer = enhancer(ReportsDetail);
-
-
-export default connect(
-    ({firebase}) => ({
-        uid: firebase.auth.uid
-    })
-)(SummaryEnhancer);
+export default ReportsDetail;
